replace all heredocs by snapshots
