+++ tests/Console/ModelsCommand/Interfaces/Test.php
@@ -46,34 +46,6 @@ final class Test extends AbstractModelsCommand
 
         $this->assertSame(0, $tester->getStatusCode());
         $this->assertEmpty($tester->getDisplay());
-
-        $expectedContent = <<<'PHP'
-<?php
-
-// @formatter:off
-/**
- * A helper file for your Eloquent Models
- * Copy the phpDocs from this file to the correct Model,
- * And remove them from this file, to prevent double declarations.
- *
- * @author Barry vd. Heuvel <barryvdh@gmail.com>
- */
-
-
-namespace Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Interfaces\Models{
-/**
- * Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\Interfaces\Models\User
- *
- * @method static \Illuminate\Database\Eloquent\Builder|User newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|User newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|User query()
- */
-	class User extends \Eloquent implements \Illuminate\Contracts\Auth\Authenticatable {}
-}
-
-
-PHP;
-
-        $this->assertSame($expectedContent, $actualContent);
+        $this->assertMatchesPhpSnapshot($actualContent);
     }
 }
